Abandonware
Fixed the scripts to actually work, but this thing is beyond repair. Why use so many rounds of encryption, a single round is just as good as 1000 in this case. Decryption requires the password and digging that out is the hard part, number of rounds is completely irrelevant. Also these scripts rename the source files out from under Xcode, so if compilation fails, the second script is never run, you are left with the obfuscated files and the originals renamed to 'bak' files.